refactor: split metadata reducer leaves
